enemy explosion and small tweaks for Jacob

+++ src/Engine/System/PathfindingSystem.hpp
@@ -17,9 +17,9 @@ public:
 	struct vecHash
 	{
 	  size_t operator()(const glm::vec3 &v) const {
-	    size_t h1 = std::hash<int>()(round(v.x));
-	    size_t h2 = std::hash<int>()(round(v.y));
-	    size_t h3 = std::hash<int>()(round(v.z));
+	    size_t h1 = std::hash<int>()(int(round(v.x)));
+	    size_t h2 = std::hash<int>()(int(round(v.y)));
+	    size_t h3 = std::hash<int>()(int(round(v.z)));
 	    return (h1 ^ (h2 << 1)) ^ h3;
 	  }
 	};
